pyproject.toml
[project]
name = "gemm-model"
version = "0.1.0"
description = "A simple perf model for dense GEMM on GPU-style hardware"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nvidia-cutlass-dsl==4.3.0.dev0",
    "torch>=2.9.0",
    "hydra-core>=1.3.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.7",
]
